added disp hokensho image

+++ reception/src/main/java/jp/chang/myclinic/reception/javafx/HokenshoListStage.java
@@ -1,5 +1,6 @@
 package jp.chang.myclinic.reception.javafx;
 
+import javafx.application.Platform;
 import javafx.scene.Scene;
 import javafx.scene.control.Hyperlink;
 import javafx.scene.control.Label;
@@ -8,15 +9,17 @@ import javafx.stage.Stage;
 import jp.chang.myclinic.client.Service;
 import jp.chang.myclinic.reception.ReceptionEnv;
 import jp.chang.myclinic.utilfx.GuiUtil;
-import okhttp3.ResponseBody;
+import jp.chang.myclinic.utilfx.HandlerFX;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.List;
+import java.util.function.Consumer;
 
 class HokenshoListStage extends Stage {
 
@@ -26,7 +29,7 @@ class HokenshoListStage extends Stage {
         String label;
         String file;
 
-        Model(String label, String file){
+        Model(String label, String file) {
             this.label = label;
             this.file = file;
         }
@@ -37,7 +40,7 @@ class HokenshoListStage extends Stage {
         VBox root = new VBox(4);
         root.getStyleClass().addAll("dialog-root", "list-hokensho-dialog");
         root.getStylesheets().add("css/Main.css");
-        if( models.size() == 0 ){
+        if (models.size() == 0) {
             root.getChildren().add(new Label("（空白）"));
         } else {
             models.forEach(model -> {
@@ -49,20 +52,35 @@ class HokenshoListStage extends Stage {
         setScene(new Scene(root));
     }
 
-    private void doOpenImage(int patientId, String file){
-        try {
-            Path saveDir = ReceptionEnv.INSTANCE.getImageSaveDir();
-            ResponseBody body = Service.api.getHokenshoCall(patientId, file)
-                    .execute().body();
-            Path savePath = Paths.get(saveDir.toString(), file);
-            try(FileOutputStream outStream = new FileOutputStream(savePath.toFile())) {
-                body.byteStream().transferTo(outStream);
-                savePath.toFile().deleteOnExit();
-            }
-        } catch (IOException e) {
-            logger.error("Failed to download file. {}", file, e);
-            GuiUtil.alertError("Failed to download file.");
+    private void doOpenImage(int patientId, String file) {
+        downloadImage(patientId, file, path -> {
+            ImageDisplayStage imageStage = new ImageDisplayStage(path);
+            imageStage.initOwner(this);
+            imageStage.show();
+        });
+    }
+
+    private void downloadImage(int patientId, String file, Consumer<Path> cb) {
+        Path saveDir = ReceptionEnv.INSTANCE.getImageSaveDir();
+        Path savePath = Paths.get(saveDir.toString(), file);
+        if (!Files.exists(savePath)) {
+            Service.api.getHokensho(patientId, file)
+                    .thenAccept(responseBody -> {
+                        try (FileOutputStream outStream = new FileOutputStream(savePath.toFile())) {
+                            responseBody.byteStream().transferTo(outStream);
+                            savePath.toFile().deleteOnExit();
+                            Platform.runLater(() -> cb.accept(savePath));
+                        } catch (IOException e) {
+                            logger.error("Failed to download file. {}", file, e);
+                            GuiUtil.alertError("Failed to download file.");
+                        }
+                    })
+                    .exceptionally(HandlerFX::exceptionally);
+        } else {
+            cb.accept(savePath);
         }
     }
+
+
 
 }
